requirements.txt
flask
flask_sqlalchemy
flask_cors
gunicorn
sqlalchemy
flask-cors
psycopg2-binary
requests
cryptography
routes
google-generativeai
